Basically the book code working with some rspec examples.  Different from the book code in that it uses both the array operators (<<, []=) and the parent pointers at once.  While the book only did one or the other not both together, and yes there are some tweaks needed to get this to work.
One learning experience from doing this is the "should make batter by instantiating subtasks" examples set an expectation that "new" be called on the subtasks.  I didn't realize that this means "new" is being stubbed and so overwriting the real "new" method.  If you don't explictly add "and_returns" to return something reasonable, then nil gets returned and the setting of parent fails because "parent=" is not a method on nil.